+++ resourcegen/names.go
@@ -18,7 +18,7 @@ var goKeywords = map[string]struct{}{
 var reservedPackages = map[string]struct{}{
 	"platform": {}, "cmd": {}, "config": {}, "database": {}, "frontend": {},
 	"internal": {}, "main": {}, "testdata": {}, "vendor": {}, "commands": {},
-	"web": {},
+	"web": {}, "product": {},
 }
 
 var reservedFields = map[string]struct{}{